+++ lib/flutter_osm_plugin.dart
@@ -3,6 +3,7 @@ library osm_flutter;
 export './src/geo_point.dart';
 export './src/marker.dart';
 export './src/road.dart';
+export './src/geo_static.dart';
 export './src/osm_flutter.dart';
 export './src/geo_point_exception.dart';
 export './src/road_exception.dart';