+++ lib/src/job_alert_dialog.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 
 class JobAlertDialog extends StatefulWidget {
   final VoidCallback callback;
+
   JobAlertDialog({this.callback});
+
   @override
   State<StatefulWidget> createState() => _JobAlertDialogState();
 }
@@ -11,7 +13,7 @@ class _JobAlertDialogState extends State<JobAlertDialog> {
   @override
   void initState() {
     super.initState();
-    Future.delayed(Duration.zero,() {
+    Future.delayed(Duration.zero, () {
       widget.callback();
     });
   }